+++ src/config/gen.ts
@@ -128,6 +128,7 @@ export const genPartialConfigFromDryInstances = async (instances: T.DryInstances
   
   const interfaceConfig = {
     admin: true,
+    choose_free_port: true,
     id: env.interfaceId,
     driver: {
       type: 'websocket',